+++ wavetorch/source.py
@@ -6,8 +6,8 @@ class WaveSource(torch.nn.Module):
     def __init__(self, x, y):
         super().__init__()
 
-        self.register_buffer('x', to_tensor(x))
-        self.register_buffer('y', to_tensor(y))
+        self.register_buffer('x', to_tensor(x, dtype=torch.int64))
+        self.register_buffer('y', to_tensor(y, dtype=torch.int64))
 
     def forward(self, Y, X, dt=1.0):
         Y[:, self.x, self.y] = Y[:, self.x, self.y] + dt**2 * X.expand_as(Y[:, self.x, self.y])
